Fixes for compiler warnings

+++ src/psconvert.c
@@ -1222,7 +1222,7 @@ int GMT_psconvert (void *V_API, int mode, void *args) {
 		if ((T = GMT_Read_Data (API, GMT_IS_TEXTSET, GMT_IS_FILE, GMT_IS_NONE, GMT_READ_NORMAL, NULL, Ctrl->L.file, NULL)) == NULL)
 			Return (GMT_RUNTIME_ERROR);
 
-		Ctrl->In.n_files = T->n_records;
+		Ctrl->In.n_files = (unsigned int)T->n_records;
 		ps_names = gmt_M_memory (GMT, NULL, T->n_records, char *);
 		for (k = 0; k < T->table[0]->segment[0]->n_rows; k++)	/* Set pointers */
 			ps_names[k] = T->table[0]->segment[0]->data[k];
